Break up secure_dir_views into four modules
